+++ Fika.Core/Networking/Models/SetHostRequest.cs
@@ -13,7 +13,7 @@ public struct SetHostRequest
     public string[] Ips;
 
     [DataMember(Name = "port")]
-    public int Port;
+    public ushort Port;
 
     [DataMember(Name = "natPunch")]
     public bool NatPunch;
@@ -21,7 +21,7 @@ public struct SetHostRequest
     [DataMember(Name = "isHeadless")]
     public bool IsHeadless;
 
-    public SetHostRequest(string[] ips, int port, bool natPunch, bool isHeadless)
+    public SetHostRequest(string[] ips, ushort port, bool natPunch, bool isHeadless)
     {
         ServerId = CoopHandler.GetServerId();
         Ips = ips;
